requirements.txt
beautifulsoup4==4.11.1
numpy==1.21.6
pandas==1.3.4
Pillow==9.2.0
plotly==5.1.0
requests==2.26.0
selenium==4.3.0
selenium_wire==4.6.3
streamlit==1.10.0
webdriver_manager==3.5.4
openpyxl==3.0.9
streamlit-aggrid